requirements.txt
flask
flask-mysqldb
python-dotenv
flask-wtf
# CSRFProtect